+++ src/app/ui/header.tsx
@@ -1,0 +1,12 @@
+import '@/styles/styles.css';
+
+export function Header() {
+  return (
+    <div className="header">
+      <h1>
+        <img src="brewfinder_logo.png" /> Brew Finder
+      </h1>
+    </div>
+  )
+}
+
