+++ src/app/ui/sidebar.tsx
@@ -1,0 +1,52 @@
+import { useState } from 'react';
+import '@/styles/styles.css';
+
+export function Sidebar() {
+  return (
+    <div id="sidebar">
+      <div id="side-sub">
+        <ZipForm />
+      </div>
+    </div>
+  )
+}
+
+function ZipForm() {
+  const [zip, setZip] = useState("");
+
+  const handleSubmit = async (event) => {
+    event.preventDefault();
+    const searchResults = await zipSearch(zip);
+    console.log('results:', searchResults);
+  }
+
+  async function zipSearch(zip) {
+    const brewDB = 'https://api.openbrewerydb.org/v1/breweries';
+    const byPostal = '?by_postal='
+    try {
+      const response = await fetch(`${brewDB}${byPostal}${zip}`);
+      if (!response.ok) {
+        throw new Error(`Response status: ${response.status}`);
+      }
+
+      const brewSet = await response.json();
+      return brewSet;
+    } catch (error) {
+      console.error(error.message);
+    }
+    return [];
+  }
+
+  return (
+    <form onSubmit={handleSubmit}>
+      <label>Zip Code:
+        <input 
+          type="text" 
+          value={zip}
+          onChange={(e) => setZip(e.target.value)}
+        />
+      </label>
+      <input type="submit" />
+    </form>
+  )
+}
